remove terrain factories change terrain layer to use terrain and strategies
