+++ sdkproject/Assets/Mapbox/Examples/3_Globe/Scripts/SpawnOnGlobeExample.cs
@@ -3,11 +3,12 @@
 	using UnityEngine;
 	using Mapbox.Unity.MeshGeneration.Factories;
 	using Mapbox.Unity.Utilities;
+	using Mapbox.Unity.MeshGeneration.Factories.TerrainStrategies;
 
 	public class SpawnOnGlobeExample : MonoBehaviour
 	{
 		[SerializeField]
-		FlatSphereTerrainFactory _globeFactory;
+		TerrainFactoryBase _globeFactory;
 
 		[SerializeField]
 		[Geocode]
@@ -25,7 +26,7 @@
 			{
 				var instance = Instantiate(_markerPrefab);
 				var location = Conversions.StringToLatLon(locationString);
-				instance.transform.position = Conversions.GeoToWorldGlobePosition(location, _globeFactory.Radius);
+				instance.transform.position = Conversions.GeoToWorldGlobePosition(location, (_globeFactory.Strategy as FlatSphereTerrainStrategy).Radius);
 				instance.transform.localScale = Vector3.one * _spawnScale;
 				instance.transform.SetParent(transform);
 			}
