--- sdkproject/Assets/Mapbox/Unity/Editor/FlatTerrainFactoryEditor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 453eec5ac1ffe4117a19cd397e6956a3
-timeCreated: 1495569574
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
