--- sdkproject/Assets/Mapbox/Unity/Editor/TerrainWithSideWallsFactoryEditor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 892db61955c75204eb4f906e6f82be51
-timeCreated: 1508723619
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
