--- sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/FlatTerrainFactory.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0a6523094c07d4ee683da5d826fce510
-timeCreated: 1495046511
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
