+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9a7ab3f3c54be5140aecaed29e7e4f65
+folderAsset: yes
+timeCreated: 1522755730
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
