+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevatedTerrainStrategy.cs
@@ -1,22 +1,14 @@
-namespace Mapbox.Unity.MeshGeneration.Factories
-{
-	using System.Collections.Generic;
-	using UnityEngine;
-	using Mapbox.Map;
-	using Mapbox.Unity.MeshGeneration.Enums;
-	using Mapbox.Unity.MeshGeneration.Data;
-	using Utils;
-	using Mapbox.Unity.Map;
-	using System;
+﻿using System.Collections.Generic;
+using UnityEngine;
+using Mapbox.Unity.MeshGeneration.Data;
+using Mapbox.Unity.Map;
+using Mapbox.Map;
+using Mapbox.Utils;
 
-	/// <summary>
-	/// Uses Mapbox Terrain api and creates terrain meshes.
-	/// </summary>
-	[CreateAssetMenu(menuName = "Mapbox/Factories/Terrain Factory - Side Walls")]
-	public class TerrainWithSideWallsFactory : AbstractTileFactory
+namespace Mapbox.Unity.MeshGeneration.Factories.TerrainStrategies
+{
+	public class ElevatedTerrainStrategy : TerrainStrategy, IElevationBasedTerrainStrategy
 	{
-		[SerializeField]
-		ElevationLayerProperties _elevationOptions = new ElevationLayerProperties();
 		Mesh _stitchTarget;
 
 		protected Dictionary<UnwrappedTileId, Mesh> _meshData;
@@ -31,26 +23,10 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 		private int _vertA, _vertB, _vertC;
 		private int _counter;
 
-		public string MapId
+		public override void OnInitialized(ElevationLayerProperties elOptions)
 		{
-			get
-			{
-				return _elevationOptions.sourceOptions.layerSource.Id;
-			}
+			base.OnInitialized(elOptions);
 
-			set
-			{
-				_elevationOptions.sourceOptions.layerSource.Id = value;
-			}
-		}
-
-		public override void SetOptions(LayerProperties options)
-		{
-			_elevationOptions = (ElevationLayerProperties)options;
-		}
-
-		internal override void OnInitialized()
-		{
 			_meshData = new Dictionary<UnwrappedTileId, Mesh>();
 			_currentTileMeshData = new MeshData();
 			_stitchTargetMeshData = new MeshData();
@@ -61,7 +37,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			_newTriangleList = new List<int>();
 		}
 
-		internal override void OnRegistered(UnityTile tile)
+		public override void OnRegistered(UnityTile tile)
 		{
 			if (_elevationOptions.unityLayerOptions.addToLayer && tile.gameObject.layer != _elevationOptions.unityLayerOptions.layerId)
 			{
@@ -71,11 +47,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			if (tile.MeshRenderer == null)
 			{
 				var renderer = tile.gameObject.AddComponent<MeshRenderer>();
-				renderer.materials = new Material[2]
-				{
-					_elevationOptions.requiredOptions.baseMaterial,
-					_elevationOptions.sideWallOptions.wallMaterial
-				};
+				renderer.material = _elevationOptions.requiredOptions.baseMaterial;
 			}
 
 			if (tile.MeshFilter == null)
@@ -89,9 +61,20 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 				tile.gameObject.AddComponent<MeshCollider>();
 			}
 
-			CreateTerrainHeight(tile);
+			GenerateTerrainMesh(tile);
 		}
 
+		public override void OnUnregistered(UnityTile tile)
+		{
+			_meshData.Remove(tile.UnwrappedTileId);
+		}
+
+		//public override void OnErrorOccurred(TileErrorEventArgs e)
+		//{
+		//	base.OnErrorOccurred(e);
+		//}
+
+		#region mesh gen
 		private void CreateBaseMesh(UnityTile tile)
 		{
 			//TODO use arrays instead of lists
@@ -115,7 +98,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						(float)(xx - tile.Rect.Center.x) * tile.TileScale,
 						0,
 						(float)(yy - tile.Rect.Center.y) * tile.TileScale));
-					_newNormalList.Add(Unity.Constants.Math.Vector3Up);
+					_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Up);
 					_newUvList.Add(new Vector2(x * 1f / (_sampleCount - 1), 1 - (y * 1f / (_sampleCount - 1))));
 				}
 			}
@@ -140,182 +123,19 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_newTriangleList.Add(vertC);
 				}
 			}
-
-			var sideVertBase = _newVertexList.Count;
-
-			var lastRow = (_sampleCount - 1) * _sampleCount;
-			var baseTriList = new List<int>();
-			for (int x = 0; x < _sampleCount; x++)
-			{
-				//side wall
-				//024
-				//135
-				_newVertexList.Add(_newVertexList[x]);
-				_newVertexList.Add(new Vector3(
-					_newVertexList[x].x,
-					-_elevationOptions.sideWallOptions.wallHeight,
-					_newVertexList[x].z));
-				_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Forward);
-				_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Forward);
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
-
-				//--- 
-
-				_newVertexList.Add(_newVertexList[x * _sampleCount]);
-				_newVertexList.Add(new Vector3(
-						_newVertexList[x * _sampleCount].x,
-					-_elevationOptions.sideWallOptions.wallHeight,
-						_newVertexList[x * _sampleCount].z));
-				_newNormalList.Add(Vector3.left);
-				_newNormalList.Add(Vector3.left);
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
-
-				//--- 
-
-				_newVertexList.Add(_newVertexList[(x + 1) * _sampleCount - 1]);
-				_newVertexList.Add(new Vector3(
-						_newVertexList[(x + 1) * _sampleCount - 1].x,
-					-_elevationOptions.sideWallOptions.wallHeight,
-						_newVertexList[(x + 1) * _sampleCount - 1].z));
-				_newNormalList.Add(Vector3.right);
-				_newNormalList.Add(Vector3.right);
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
-
-				//--- 
-
-				_newVertexList.Add(_newVertexList[lastRow + x]);
-				_newVertexList.Add(new Vector3(
-					_newVertexList[lastRow + x].x,
-					-_elevationOptions.sideWallOptions.wallHeight,
-					_newVertexList[lastRow + x].z));
-				_newNormalList.Add(Vector3.back);
-				_newNormalList.Add(Vector3.back);
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
-				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
-
-				if (x > 0)
-				{
-					baseTriList.Add(sideVertBase + 8 * x);
-					baseTriList.Add(sideVertBase + 8 * x - 8);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1);
-
-					baseTriList.Add(sideVertBase + 8 * x);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1);
-					baseTriList.Add(sideVertBase + 8 * x + 1);
-
-					//---
-
-					baseTriList.Add(sideVertBase + 8 * x + 2);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 2);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 2);
-
-					baseTriList.Add(sideVertBase + 8 * x + 2);
-					baseTriList.Add(sideVertBase + 8 * x + 1 + 2);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 2);
-
-					//---
-
-					baseTriList.Add(sideVertBase + 8 * x + 4);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 4);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 4);
-
-					baseTriList.Add(sideVertBase + 8 * x + 4);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 4);
-					baseTriList.Add(sideVertBase + 8 * x + 1 + 4);
-
-					//---
-
-					baseTriList.Add(sideVertBase + 8 * x + 6);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 6);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 6);
-
-					baseTriList.Add(sideVertBase + 8 * x + 6);
-					baseTriList.Add(sideVertBase + 8 * x + 1 + 6);
-					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 6);
-				}
-			}
-
-
 			var mesh = tile.MeshFilter.mesh;
 			mesh.SetVertices(_newVertexList);
 			mesh.SetNormals(_newNormalList);
 			mesh.SetUVs(0, _newUvList);
-			mesh.subMeshCount = 2;
 			mesh.SetTriangles(_newTriangleList, 0);
-			mesh.SetTriangles(baseTriList, 1);
 		}
 
-		internal override void OnUnregistered(UnityTile tile)
-		{
-			_meshData.Remove(tile.UnwrappedTileId);
-		}
-
-		/// <summary>
-		/// Creates the non-flat terrain using a height multiplier
-		/// </summary>
-		/// <param name="tile"></param>
-		// <param name="heightMultiplier">Multiplier for queried height value</param>
-		private void CreateTerrainHeight(UnityTile tile)
-		{
-			tile.HeightDataState = TilePropertyState.Loading;
-			var pngRasterTile = new RawPngRasterTile();
-
-			tile.AddTile(pngRasterTile);
-			Progress++;
-
-			pngRasterTile.Initialize(_fileSource, tile.CanonicalTileId, MapId, () =>
-			{
-				if (tile == null)
-				{
-					return;
-				}
-
-				if (pngRasterTile.HasError)
-				{
-					OnErrorOccurred(new TileErrorEventArgs(tile.CanonicalTileId, pngRasterTile.GetType(), tile, pngRasterTile.Exceptions));
-					tile.HeightDataState = TilePropertyState.Error;
-
-					// Handle missing elevation from server (404)!
-					// TODO: optimize this search!
-					if (pngRasterTile.ExceptionsAsString.Contains("404"))
-					{
-						ResetToFlatMesh(tile);
-					}
-					Progress--;
-					return;
-				}
-
-				tile.SetHeightData(pngRasterTile.Data, _elevationOptions.requiredOptions.exaggerationFactor);
-				GenerateTerrainMesh(tile);
-				Progress--;
-			});
-		}
-
-		/// <summary>
-		/// Method to be called when a tile error has occurred.
-		/// </summary>
-		/// <param name="e"><see cref="T:Mapbox.Map.TileErrorEventArgs"/> instance/</param>
-		protected override void OnErrorOccurred(TileErrorEventArgs e)
-		{
-			base.OnErrorOccurred(e);
-		}
-
-		/// <summary>
-		/// Creates the non-flat terrain mesh, using a grid by defined resolution (_sampleCount). Vertex order goes right & up. Normals are calculated manually and UV map is fitted/stretched 1-1.
-		/// Any additional scripts or logic, like MeshCollider or setting layer, can be done here.
-		/// </summary>
-		/// <param name="tile"></param>
-		// <param name="heightMultiplier">Multiplier for queried height value</param>
 		private void GenerateTerrainMesh(UnityTile tile)
 		{
 			tile.MeshFilter.mesh.GetVertices(_currentTileMeshData.Vertices);
 			tile.MeshFilter.mesh.GetNormals(_currentTileMeshData.Normals);
 
 			var _sampleCount = _elevationOptions.modificationOptions.sampleCount;
-			int sideStart = _sampleCount * _sampleCount;
 			for (float y = 0; y < _sampleCount; y++)
 			{
 				for (float x = 0; x < _sampleCount; x++)
@@ -324,25 +144,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						_currentTileMeshData.Vertices[(int)(y * _sampleCount + x)].x,
 						tile.QueryHeightData(x / (_sampleCount - 1), 1 - y / (_sampleCount - 1)),
 						_currentTileMeshData.Vertices[(int)(y * _sampleCount + x)].z);
-					_currentTileMeshData.Normals[(int)(y * _sampleCount + x)] = Unity.Constants.Math.Vector3Zero;
-
-					if (y == 0)
-					{
-						_currentTileMeshData.Vertices[(int)(sideStart + 8 * x)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
-					}
-					else if (y == _sampleCount - 1)
-					{
-						_currentTileMeshData.Vertices[(int)(sideStart + 8 * x + 6)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
-					}
-
-					if (x == 0)
-					{
-						_currentTileMeshData.Vertices[(int)(sideStart + 8 * y + 2)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
-					}
-					else if (x == _sampleCount - 1)
-					{
-						_currentTileMeshData.Vertices[(int)(sideStart + 8 * y + 4)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
-					}
+					_currentTileMeshData.Normals[(int)(y * _sampleCount + x)] = Mapbox.Unity.Constants.Math.Vector3Zero;
 				}
 			}
 
@@ -404,7 +206,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_currentTileMeshData.Vertices[i].x,
 					0,
 					_currentTileMeshData.Vertices[i].z);
-				_currentTileMeshData.Normals[i] = Unity.Constants.Math.Vector3Up;
+				_currentTileMeshData.Normals[i] = Mapbox.Unity.Constants.Math.Vector3Up;
 			}
 
 			tile.MeshFilter.mesh.SetVertices(_currentTileMeshData.Vertices);
@@ -416,18 +218,19 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 		/// <summary>
 		/// Checkes all neighbours of the given tile and stitches the edges to achieve a smooth mesh surface.
 		/// </summary>
-		/// <param name="tileId"></param>
+		/// <param name="tileId">UnwrappedTileId of the tile being processed.</param>
 		/// <param name="mesh"></param>
 		private void FixStitches(UnwrappedTileId tileId, MeshData mesh)
 		{
 			var _sampleCount = _elevationOptions.modificationOptions.sampleCount;
-			var meshVertCount = _sampleCount * _sampleCount;
+			var meshVertCount = mesh.Vertices.Count;
 			_stitchTarget = null;
 			_meshData.TryGetValue(tileId.North, out _stitchTarget);
 			if (_stitchTarget != null)
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
+
 				for (int i = 0; i < _sampleCount; i++)
 				{
 					//just snapping the y because vertex pos is relative and we'll have to do tile pos + vertex pos for x&z otherwise
@@ -435,7 +238,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i].x,
 						_stitchTargetMeshData.Vertices[meshVertCount - _sampleCount + i].y,
 						mesh.Vertices[i].z);
-					mesh.Vertices[meshVertCount + (8 * i)] = mesh.Vertices[i];
 
 					mesh.Normals[i] = new Vector3(_stitchTargetMeshData.Normals[meshVertCount - _sampleCount + i].x,
 						_stitchTargetMeshData.Normals[meshVertCount - _sampleCount + i].y,
@@ -449,13 +251,13 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
+
 				for (int i = 0; i < _sampleCount; i++)
 				{
 					mesh.Vertices[meshVertCount - _sampleCount + i] = new Vector3(
 						mesh.Vertices[meshVertCount - _sampleCount + i].x,
 						_stitchTargetMeshData.Vertices[i].y,
 						mesh.Vertices[meshVertCount - _sampleCount + i].z);
-					mesh.Vertices[meshVertCount + 6 + (8 * i)] = mesh.Vertices[meshVertCount - _sampleCount + i];
 
 					mesh.Normals[meshVertCount - _sampleCount + i] = new Vector3(
 						_stitchTargetMeshData.Normals[i].x,
@@ -477,7 +279,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i * _sampleCount].x,
 						_stitchTargetMeshData.Vertices[i * _sampleCount + _sampleCount - 1].y,
 						mesh.Vertices[i * _sampleCount].z);
-					mesh.Vertices[meshVertCount + 2 + (8 * i)] = mesh.Vertices[i * _sampleCount];
 
 					mesh.Normals[i * _sampleCount] = new Vector3(
 						_stitchTargetMeshData.Normals[i * _sampleCount + _sampleCount - 1].x,
@@ -500,7 +301,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i * _sampleCount + _sampleCount - 1].x,
 						_stitchTargetMeshData.Vertices[i * _sampleCount].y,
 						mesh.Vertices[i * _sampleCount + _sampleCount - 1].z);
-					mesh.Vertices[meshVertCount + 4 + (8 * i)] = mesh.Vertices[i * _sampleCount + _sampleCount - 1];
 
 					mesh.Normals[i * _sampleCount + _sampleCount - 1] = new Vector3(
 						_stitchTargetMeshData.Normals[i * _sampleCount].x,
@@ -573,6 +373,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
+
 				mesh.Vertices[meshVertCount - 1] = new Vector3(
 					mesh.Vertices[meshVertCount - 1].x,
 					_stitchTargetMeshData.Vertices[0].y,
@@ -584,5 +385,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_stitchTargetMeshData.Normals[0].z);
 			}
 		}
+		#endregion
 	}
 }
