+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevatedTerrainStrategy.cs.meta
@@ -1,12 +1,13 @@
 fileFormatVersion: 2
-guid: 5632e34628bc2694f867c3f3c77bd796
-timeCreated: 1481718836
-licenseType: Pro
+guid: ea9b70aa9455c614f8701116ede4599d
+timeCreated: 1522755494
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 2ac69ccfbca692443a02f68e1a689694, type: 3}
+  icon: {instanceID: 0}
   userData: 
   assetBundleName: 
   assetBundleVariant:
