+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevatedTerrainWithSidesStrategy.cs
@@ -1,28 +1,14 @@
-namespace Mapbox.Unity.MeshGeneration.Factories
+﻿using System.Collections.Generic;
+using UnityEngine;
+using Mapbox.Unity.MeshGeneration.Data;
+using Mapbox.Unity.Map;
+using Mapbox.Map;
+using Mapbox.Utils;
+
+namespace Mapbox.Unity.MeshGeneration.Factories.TerrainStrategies
 {
-	using System.Collections.Generic;
-	using UnityEngine;
-	using Mapbox.Map;
-	using Mapbox.Unity.MeshGeneration.Enums;
-	using Mapbox.Unity.MeshGeneration.Data;
-	using Utils;
-	using System;
-	using Mapbox.Unity.Map;
-
-	public enum MapIdType
+	public class ElevatedTerrainWithSidesStrategy : TerrainStrategy, IElevationBasedTerrainStrategy
 	{
-		Standard,
-		Custom
-	}
-
-	/// <summary>
-	/// Uses Mapbox Terrain api and creates terrain meshes.
-	/// </summary>
-	[CreateAssetMenu(menuName = "Mapbox/Factories/Terrain Factory")]
-	public class TerrainFactory : AbstractTileFactory
-	{
-		[SerializeField]
-		ElevationLayerProperties _elevationOptions = new ElevationLayerProperties();
 		Mesh _stitchTarget;
 
 		protected Dictionary<UnwrappedTileId, Mesh> _meshData;
@@ -37,43 +23,10 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 		private int _vertA, _vertB, _vertC;
 		private int _counter;
 
-		//public string MapId
-		//{
-		//	get
-		//	{
-		//		return _elevationOptions.sourceOptions.Id;
-		//	}
-
-		//	set
-		//	{
-		//		_elevationOptions.sourceOptions.Id = value;
-		//	}
-		//}
-
-		public override void SetOptions(LayerProperties options)
+		public override void OnInitialized(ElevationLayerProperties elOptions)
 		{
-			_elevationOptions = (ElevationLayerProperties)options;
-		}
+			_elevationOptions = elOptions;
 
-		/// <summary>
-		/// Clears the mesh data and re-runs the terrain creation procedure using current settings. Clearing the old mesh data is important as terrain stitching function checks if the data exists or not.
-		/// </summary>
-		// TODO: come back to this
-		//public override void Update()
-		//{
-		//    base.Update();
-		//    foreach (var tile in _tiles.Values)
-		//    {
-		//        tile.MeshData = null;
-		//    }
-		//    foreach (var tile in _tiles.Values)
-		//    {
-		//        Run(tile);
-		//    }
-		//}
-
-		internal override void OnInitialized()
-		{
 			_meshData = new Dictionary<UnwrappedTileId, Mesh>();
 			_currentTileMeshData = new MeshData();
 			_stitchTargetMeshData = new MeshData();
@@ -84,7 +37,8 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			_newTriangleList = new List<int>();
 		}
 
-		internal override void OnRegistered(UnityTile tile)
+
+		public override void OnRegistered(UnityTile tile)
 		{
 			if (_elevationOptions.unityLayerOptions.addToLayer && tile.gameObject.layer != _elevationOptions.unityLayerOptions.layerId)
 			{
@@ -94,7 +48,11 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			if (tile.MeshRenderer == null)
 			{
 				var renderer = tile.gameObject.AddComponent<MeshRenderer>();
-				renderer.material = _elevationOptions.requiredOptions.baseMaterial;
+				renderer.materials = new Material[2]
+				{
+					_elevationOptions.requiredOptions.baseMaterial,
+					_elevationOptions.sideWallOptions.wallMaterial
+				};
 			}
 
 			if (tile.MeshFilter == null)
@@ -108,7 +66,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 				tile.gameObject.AddComponent<MeshCollider>();
 			}
 
-			CreateTerrainHeight(tile);
+			GenerateTerrainMesh(tile);
 		}
 
 		private void CreateBaseMesh(UnityTile tile)
@@ -134,7 +92,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						(float)(xx - tile.Rect.Center.x) * tile.TileScale,
 						0,
 						(float)(yy - tile.Rect.Center.y) * tile.TileScale));
-					_newNormalList.Add(Unity.Constants.Math.Vector3Up);
+					_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Up);
 					_newUvList.Add(new Vector2(x * 1f / (_sampleCount - 1), 1 - (y * 1f / (_sampleCount - 1))));
 				}
 			}
@@ -159,66 +117,117 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_newTriangleList.Add(vertC);
 				}
 			}
+
+			var sideVertBase = _newVertexList.Count;
+
+			var lastRow = (_sampleCount - 1) * _sampleCount;
+			var baseTriList = new List<int>();
+			for (int x = 0; x < _sampleCount; x++)
+			{
+				//side wall
+				//024
+				//135
+				_newVertexList.Add(_newVertexList[x]);
+				_newVertexList.Add(new Vector3(
+					_newVertexList[x].x,
+					-_elevationOptions.sideWallOptions.wallHeight,
+					_newVertexList[x].z));
+				_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Forward);
+				_newNormalList.Add(Mapbox.Unity.Constants.Math.Vector3Forward);
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
+
+				//--- 
+
+				_newVertexList.Add(_newVertexList[x * _sampleCount]);
+				_newVertexList.Add(new Vector3(
+						_newVertexList[x * _sampleCount].x,
+					-_elevationOptions.sideWallOptions.wallHeight,
+						_newVertexList[x * _sampleCount].z));
+				_newNormalList.Add(Vector3.left);
+				_newNormalList.Add(Vector3.left);
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
+
+				//--- 
+
+				_newVertexList.Add(_newVertexList[(x + 1) * _sampleCount - 1]);
+				_newVertexList.Add(new Vector3(
+						_newVertexList[(x + 1) * _sampleCount - 1].x,
+					-_elevationOptions.sideWallOptions.wallHeight,
+						_newVertexList[(x + 1) * _sampleCount - 1].z));
+				_newNormalList.Add(Vector3.right);
+				_newNormalList.Add(Vector3.right);
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
+
+				//--- 
+
+				_newVertexList.Add(_newVertexList[lastRow + x]);
+				_newVertexList.Add(new Vector3(
+					_newVertexList[lastRow + x].x,
+					-_elevationOptions.sideWallOptions.wallHeight,
+					_newVertexList[lastRow + x].z));
+				_newNormalList.Add(Vector3.back);
+				_newNormalList.Add(Vector3.back);
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 1));
+				_newUvList.Add(new Vector2(_newUvList[x * _sampleCount].y, 0));
+
+				if (x > 0)
+				{
+					baseTriList.Add(sideVertBase + 8 * x);
+					baseTriList.Add(sideVertBase + 8 * x - 8);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1);
+
+					baseTriList.Add(sideVertBase + 8 * x);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1);
+					baseTriList.Add(sideVertBase + 8 * x + 1);
+
+					//---
+
+					baseTriList.Add(sideVertBase + 8 * x + 2);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 2);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 2);
+
+					baseTriList.Add(sideVertBase + 8 * x + 2);
+					baseTriList.Add(sideVertBase + 8 * x + 1 + 2);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 2);
+
+					//---
+
+					baseTriList.Add(sideVertBase + 8 * x + 4);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 4);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 4);
+
+					baseTriList.Add(sideVertBase + 8 * x + 4);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 4);
+					baseTriList.Add(sideVertBase + 8 * x + 1 + 4);
+
+					//---
+
+					baseTriList.Add(sideVertBase + 8 * x + 6);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 6);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 6);
+
+					baseTriList.Add(sideVertBase + 8 * x + 6);
+					baseTriList.Add(sideVertBase + 8 * x + 1 + 6);
+					baseTriList.Add(sideVertBase + 8 * x - 8 + 1 + 6);
+				}
+			}
+
+
 			var mesh = tile.MeshFilter.mesh;
 			mesh.SetVertices(_newVertexList);
 			mesh.SetNormals(_newNormalList);
 			mesh.SetUVs(0, _newUvList);
+			mesh.subMeshCount = 2;
 			mesh.SetTriangles(_newTriangleList, 0);
+			mesh.SetTriangles(baseTriList, 1);
 		}
 
-		internal override void OnUnregistered(UnityTile tile)
+		public override void OnUnregistered(UnityTile tile)
 		{
 			_meshData.Remove(tile.UnwrappedTileId);
-		}
-
-		/// <summary>
-		/// Creates the non-flat terrain using a height multiplier
-		/// </summary>
-		/// <param name="tile">Tile.</param>
-		private void CreateTerrainHeight(UnityTile tile)
-		{
-			tile.HeightDataState = TilePropertyState.Loading;
-			var pngRasterTile = new RawPngRasterTile();
-
-			tile.AddTile(pngRasterTile);
-			Progress++;
-
-			pngRasterTile.Initialize(_fileSource, tile.CanonicalTileId, _elevationOptions.sourceOptions.Id, () =>
-			{
-				if (tile == null)
-				{
-					Progress--;
-					return;
-				}
-
-				if (pngRasterTile.HasError)
-				{
-					OnErrorOccurred(new TileErrorEventArgs(tile.CanonicalTileId, pngRasterTile.GetType(), tile, pngRasterTile.Exceptions));
-					tile.HeightDataState = TilePropertyState.Error;
-
-					// Handle missing elevation from server (404)!
-					// TODO: optimize this search!
-					if (pngRasterTile.ExceptionsAsString.Contains("404"))
-					{
-						ResetToFlatMesh(tile);
-					}
-					Progress--;
-					return;
-				}
-
-				tile.SetHeightData(pngRasterTile.Data, _elevationOptions.requiredOptions.exaggerationFactor, _elevationOptions.modificationOptions.useRelativeHeight);
-				GenerateTerrainMesh(tile);
-				Progress--;
-			});
-		}
-
-		/// <summary>
-		/// Method to be called when a tile error has occurred.
-		/// </summary>
-		/// <param name="e"><see cref="T:Mapbox.Map.TileErrorEventArgs"/> instance/</param>
-		protected override void OnErrorOccurred(TileErrorEventArgs e)
-		{
-			base.OnErrorOccurred(e);
 		}
 
 		/// <summary>
@@ -226,12 +235,14 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 		/// Any additional scripts or logic, like MeshCollider or setting layer, can be done here.
 		/// </summary>
 		/// <param name="tile"></param>
+		// <param name="heightMultiplier">Multiplier for queried height value</param>
 		private void GenerateTerrainMesh(UnityTile tile)
 		{
 			tile.MeshFilter.mesh.GetVertices(_currentTileMeshData.Vertices);
 			tile.MeshFilter.mesh.GetNormals(_currentTileMeshData.Normals);
 
 			var _sampleCount = _elevationOptions.modificationOptions.sampleCount;
+			int sideStart = _sampleCount * _sampleCount;
 			for (float y = 0; y < _sampleCount; y++)
 			{
 				for (float x = 0; x < _sampleCount; x++)
@@ -240,7 +251,25 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						_currentTileMeshData.Vertices[(int)(y * _sampleCount + x)].x,
 						tile.QueryHeightData(x / (_sampleCount - 1), 1 - y / (_sampleCount - 1)),
 						_currentTileMeshData.Vertices[(int)(y * _sampleCount + x)].z);
-					_currentTileMeshData.Normals[(int)(y * _sampleCount + x)] = Unity.Constants.Math.Vector3Zero;
+					_currentTileMeshData.Normals[(int)(y * _sampleCount + x)] = Mapbox.Unity.Constants.Math.Vector3Zero;
+
+					if (y == 0)
+					{
+						_currentTileMeshData.Vertices[(int)(sideStart + 8 * x)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
+					}
+					else if (y == _sampleCount - 1)
+					{
+						_currentTileMeshData.Vertices[(int)(sideStart + 8 * x + 6)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
+					}
+
+					if (x == 0)
+					{
+						_currentTileMeshData.Vertices[(int)(sideStart + 8 * y + 2)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
+					}
+					else if (x == _sampleCount - 1)
+					{
+						_currentTileMeshData.Vertices[(int)(sideStart + 8 * y + 4)] = _currentTileMeshData.Vertices[(int)(y * _sampleCount + x)];
+					}
 				}
 			}
 
@@ -302,7 +331,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_currentTileMeshData.Vertices[i].x,
 					0,
 					_currentTileMeshData.Vertices[i].z);
-				_currentTileMeshData.Normals[i] = Unity.Constants.Math.Vector3Up;
+				_currentTileMeshData.Normals[i] = Mapbox.Unity.Constants.Math.Vector3Up;
 			}
 
 			tile.MeshFilter.mesh.SetVertices(_currentTileMeshData.Vertices);
@@ -314,19 +343,18 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 		/// <summary>
 		/// Checkes all neighbours of the given tile and stitches the edges to achieve a smooth mesh surface.
 		/// </summary>
-		/// <param name="tileId">UnwrappedTileId of the tile being processed.</param>
+		/// <param name="tileId"></param>
 		/// <param name="mesh"></param>
 		private void FixStitches(UnwrappedTileId tileId, MeshData mesh)
 		{
 			var _sampleCount = _elevationOptions.modificationOptions.sampleCount;
-			var meshVertCount = mesh.Vertices.Count;
+			var meshVertCount = _sampleCount * _sampleCount;
 			_stitchTarget = null;
 			_meshData.TryGetValue(tileId.North, out _stitchTarget);
 			if (_stitchTarget != null)
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
-
 				for (int i = 0; i < _sampleCount; i++)
 				{
 					//just snapping the y because vertex pos is relative and we'll have to do tile pos + vertex pos for x&z otherwise
@@ -334,6 +362,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i].x,
 						_stitchTargetMeshData.Vertices[meshVertCount - _sampleCount + i].y,
 						mesh.Vertices[i].z);
+					mesh.Vertices[meshVertCount + (8 * i)] = mesh.Vertices[i];
 
 					mesh.Normals[i] = new Vector3(_stitchTargetMeshData.Normals[meshVertCount - _sampleCount + i].x,
 						_stitchTargetMeshData.Normals[meshVertCount - _sampleCount + i].y,
@@ -347,13 +376,13 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
-
 				for (int i = 0; i < _sampleCount; i++)
 				{
 					mesh.Vertices[meshVertCount - _sampleCount + i] = new Vector3(
 						mesh.Vertices[meshVertCount - _sampleCount + i].x,
 						_stitchTargetMeshData.Vertices[i].y,
 						mesh.Vertices[meshVertCount - _sampleCount + i].z);
+					mesh.Vertices[meshVertCount + 6 + (8 * i)] = mesh.Vertices[meshVertCount - _sampleCount + i];
 
 					mesh.Normals[meshVertCount - _sampleCount + i] = new Vector3(
 						_stitchTargetMeshData.Normals[i].x,
@@ -375,6 +404,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i * _sampleCount].x,
 						_stitchTargetMeshData.Vertices[i * _sampleCount + _sampleCount - 1].y,
 						mesh.Vertices[i * _sampleCount].z);
+					mesh.Vertices[meshVertCount + 2 + (8 * i)] = mesh.Vertices[i * _sampleCount];
 
 					mesh.Normals[i * _sampleCount] = new Vector3(
 						_stitchTargetMeshData.Normals[i * _sampleCount + _sampleCount - 1].x,
@@ -397,6 +427,7 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 						mesh.Vertices[i * _sampleCount + _sampleCount - 1].x,
 						_stitchTargetMeshData.Vertices[i * _sampleCount].y,
 						mesh.Vertices[i * _sampleCount + _sampleCount - 1].z);
+					mesh.Vertices[meshVertCount + 4 + (8 * i)] = mesh.Vertices[i * _sampleCount + _sampleCount - 1];
 
 					mesh.Normals[i * _sampleCount + _sampleCount - 1] = new Vector3(
 						_stitchTargetMeshData.Normals[i * _sampleCount].x,
@@ -469,7 +500,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 			{
 				_stitchTarget.GetVertices(_stitchTargetMeshData.Vertices);
 				_stitchTarget.GetNormals(_stitchTargetMeshData.Normals);
-
 				mesh.Vertices[meshVertCount - 1] = new Vector3(
 					mesh.Vertices[meshVertCount - 1].x,
 					_stitchTargetMeshData.Vertices[0].y,
@@ -481,5 +511,6 @@ namespace Mapbox.Unity.MeshGeneration.Factories
 					_stitchTargetMeshData.Normals[0].z);
 			}
 		}
+
 	}
 }
