+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevatedTerrainWithSidesStrategy.cs.meta
@@ -1,12 +1,13 @@
 fileFormatVersion: 2
-guid: 8338f26893d62c1439f387800a8e1c9a
-timeCreated: 1481718836
-licenseType: Pro
+guid: 3d67aa1b17136294b92c1a5f4d9d867a
+timeCreated: 1522755493
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
-  icon: {fileID: 2800000, guid: 2ac69ccfbca692443a02f68e1a689694, type: 3}
+  icon: {instanceID: 0}
   userData: 
   assetBundleName: 
   assetBundleVariant:
