+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevationBasedTerrainStrategy.cs
@@ -1,0 +1,4 @@
+﻿public interface IElevationBasedTerrainStrategy
+{
+
+}
