+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/ElevationBasedTerrainStrategy.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 859bc316b9b734bffaf7288cb86ffc0c
-timeCreated: 1501082693
-licenseType: Pro
+guid: 0629e9d95db9602439cbd191b7c35260
+timeCreated: 1522756350
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
