+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/FlatSphereTerrainStrategy.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 9ffef98732ebca048a22c90f6809321c
-timeCreated: 1500160220
-licenseType: Pro
+guid: 830e93932f0e72840a1f6d588dce1a4a
+timeCreated: 1522755493
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
