+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/FlatTerrainStrategy.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: f20150ceca635c44dab37cfe0460aa36
-timeCreated: 1490923136
-licenseType: Pro
+guid: dc77d6f0c5be2cb49bf8014f58feeb43
+timeCreated: 1522755493
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
