+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/LowPolyTerrainStrategy.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 94b2783afb8700b4a857a77b0b87f846
-timeCreated: 1495569574
-licenseType: Pro
+guid: 2e1d75acbfaded5459635c6030b0f915
+timeCreated: 1522755493
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
