+++ sdkproject/Assets/Mapbox/Unity/MeshGeneration/Factories/TerrainStrategies/TerrainStrategy.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 1ae8d3eabc0344593b4b8d02c19bf67d
-timeCreated: 1501082693
-licenseType: Pro
+guid: 33b4040c296ec314793ac7182cc3e81c
+timeCreated: 1522755493
+licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
